Modify test file suffix
